Move Cluster to separate module
