+++ fiaas_skipper/deploy/__init__.py
@@ -4,8 +4,6 @@ from __future__ import absolute_import
 
 from .channel import ReleaseChannelFactory
 from .crd import CrdDeployer, CrdBootstrapper
-from .deploy import Cluster, DeploymentConfigStatus
 from .tpr import TprDeployer, TprBootstrapper
 
-__all__ = ["TprDeployer", "TprBootstrapper", "CrdDeployer", "CrdBootstrapper", "Cluster",
-           "ReleaseChannelFactory", "DeploymentConfigStatus"]
+__all__ = ["TprDeployer", "TprBootstrapper", "CrdDeployer", "CrdBootstrapper", "ReleaseChannelFactory"]
